# Refactor Tx Search Aggregation