- Replace the previous `intersect` call, which was called at each query condition, with a map intersection.
- Replace `fmt.Sprintf` with `string()`

closes: #3076

----

## Benchmarks 

### Old

```
goos: darwin
goarch: amd64
pkg: github.com/tendermint/tendermint/state/txindex/kv
BenchmarkTxSearch-4   	     200	 103641206 ns/op	 7998416 B/op	   71171 allocs/op
PASS
ok  	github.com/tendermint/tendermint/state/txindex/kv	26.019s
```

### New

```
goos: darwin
goarch: amd64
pkg: github.com/tendermint/tendermint/state/txindex/kv
BenchmarkTxSearch-4   	    1000	  38615024 ns/op	13515226 B/op	  166460 allocs/op
PASS
ok  	github.com/tendermint/tendermint/state/txindex/kv	53.618s
```

**~62% performance improvement**

## Profile

### New

![pprof002](https://user-images.githubusercontent.com/5265989/62217613-ef13a400-b378-11e9-9f22-0e7c3f1de20b.png)

At this point, it seems the only bottleneck is the DB itself.

----

* [x] Referenced an issue explaining the need for the change
* [ ] Updated all relevant documentation in docs
* [x] Updated all code comments where relevant
* [x] Wrote tests
* [x] Updated CHANGELOG_PENDING.md